pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "digital-twin-tumor"
version = "0.1.0"
description = "Digital Twin for Tumor Response Assessment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Digital Twin Tumor Team"},
]
keywords = ["medical-imaging", "digital-twin", "tumor", "recist", "medgemma", "oncology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "scikit-image>=0.22",
    "pydicom>=2.4",
    "nibabel>=5.0",
    "SimpleITK>=2.3",
    "networkx>=3.2",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "gradio>=5.0",
    "pydantic>=2.5",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.1",
    "transformers>=4.50",
    "accelerate>=0.25",
    "bitsandbytes>=0.42",
]
storage = [
    "psycopg[binary]>=3.1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
digital-twin-tumor = "digital_twin_tumor.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
